requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
crispy-bootstrap5==2025.4
distro==1.9.0
Django==5.2
django-crispy-forms==2.4
djangorestframework==3.15.2
et_xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mypy_extensions==1.1.0
numpy==2.2.4
openai==1.63.1
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pydantic==2.10.6
pydantic_core==2.27.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
PySide2==5.13.2
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
rdkit==2024.9.6
requests==2.32.3
shiboken2==5.13.2
shiboken6==6.8.2.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
